feat(ffe-system-message-react): Reimplement and improve
This commit contributes a new implementation of the
`ffe-system-message-react` package.

The change removes an optional `animationLengthMs` property, which
now is set to 200ms. It's not breaking to leave it in place, but it
does no longer affect anything.

This new implementation brings a few new nice-to-haves to the
`ffe-system-message-react` components:

- Props that are not listed are passed to the outer component
- The system message is now laid out using ffe-grid
- The close button is now a `<button />`, improving a11y
- The test coverage is 100 %, ensuring a stable implementation with
few bugs